🗻: Complete Oauth Link Components
Completes the oauth link components for our Twitch and Discord
integrations.

Currently they send an authorize request to their respective service,
which then returns back an auth code after the user has successfully
logged in.

We still need to send out a request to the backend to handle the
code we receive from Discord/Twitch.